hw9-bookshelf-spring-jdbc: refactored query service folder
