+++ src/main/java/com/otus/bookstore/dao/impl/BookDaoJdbc.java
@@ -2,7 +2,7 @@ package com.otus.bookstore.dao.impl;
 
 import com.otus.bookstore.dao.BookDao;
 import com.otus.bookstore.model.Book;
-import com.otus.bookstore.service.book.*;
+import com.otus.bookstore.service.query.book.*;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
