+++ src/main/java/com/otus/bookstore/service/query/author/AuthorSelectAllQuery.java
@@ -1,20 +1,17 @@
-package com.otus.bookstore.service.author;
+package com.otus.bookstore.service.query.author;
 
 import com.otus.bookstore.model.Author;
-import org.springframework.jdbc.core.SqlParameter;
 import org.springframework.jdbc.object.MappingSqlQuery;
 
 import javax.sql.DataSource;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Types;
 
-public class AuthorSelectByIdQuery extends MappingSqlQuery<Author> {
-    private static final String SQL_SELECT_AUTHOR_BY_ID = "SELECT * FROM author WHERE id = :id";
+public class AuthorSelectAllQuery extends MappingSqlQuery<Author> {
+    private static final String SQL_SELECT_ALL_AUTHORS = "SELECT * FROM author";
 
-    public AuthorSelectByIdQuery(DataSource dataSource) {
-        super(dataSource, SQL_SELECT_AUTHOR_BY_ID);
-        declareParameter(new SqlParameter("id", Types.INTEGER));
+    public AuthorSelectAllQuery(DataSource dataSource) {
+        super(dataSource, SQL_SELECT_ALL_AUTHORS);
     }
 
     @Override
@@ -25,3 +22,4 @@ public class AuthorSelectByIdQuery extends MappingSqlQuery<Author> {
         return new Author(id, name, email);
     }
 }
+
