+++ src/main/java/com/otus/bookstore/service/query/book/BookSelectByIdQuery.java
@@ -1,4 +1,4 @@
-package com.otus.bookstore.service.book;
+package com.otus.bookstore.service.query.book;
 
 import com.otus.bookstore.model.Book;
 import org.springframework.jdbc.core.SqlParameter;
