+++ src/main/java/com/otus/bookstore/service/query/genre/GenreUpdateQuery.java
@@ -1,4 +1,4 @@
-package com.otus.bookstore.service.genre;
+package com.otus.bookstore.service.query.genre;
 
 import org.springframework.jdbc.core.SqlParameter;
 import org.springframework.jdbc.object.SqlUpdate;
